+++ R/box_module_usage_helper_functions.R
@@ -39,7 +39,7 @@ get_attached_modules <- function(xml) {
   nested_list <- get_module_exports(attached_modules$text)
   # normalize module names
   attached_modules$text <- basename(attached_modules$text)
-  names(nested_list) <- basename(names(nested_list))
+  names(nested_list) <- gsub("`", "", basename(names(nested_list)))
 
   whole_module_imports <- "
   /child::*[