+++ tests/testthat/mod/path/to/01_module.R
@@ -1,0 +1,9 @@
+#' @export
+a_fun_a <- function() {
+  "exported a"
+}
+
+#' @export
+`01_fun` <- function() {
+  "exported a"
+}