+++ NEWS.md
@@ -1,6 +1,8 @@
 # development version
 
 * Fix for false-positive lint on a list-name function call of a function argument. (#131)
+* Support for destructure operator `%<-%`
+* Support for non-syntactic names for objects, functions, module names (#147, #151)
 
 # box.linters 0.10.5